requirements.txt
discord.py>=2.3.2
yt-dlp>=2024.10.0
python-dotenv>=1.0.1

